feat: introduce ReflectionMode attribute and enhance execution mode handling
