+++ TUnit.PublicAPI/Tests.Core_Library_Has_No_API_Changes.DotNet9_0.verified.txt
@@ -1104,6 +1104,11 @@ namespace
         public static ..IPropertySource? GetSource( type) { }
         public static void Register( type, ..IPropertySource source) { }
     }
+    [(.Assembly, AllowMultiple=false, Inherited=false)]
+    public sealed class ReflectionModeAttribute : 
+    {
+        public ReflectionModeAttribute() { }
+    }
     [(.Assembly | .Class | .Method)]
     public sealed class RepeatAttribute : .TUnitAttribute, .IScopedAttribute
     {
